+++ src/Symfony/Component/Form/Extension/Core/DataTransformer/MoneyToLocalizedStringTransformer.php
@@ -45,7 +45,7 @@ class MoneyToLocalizedStringTransformer extends NumberToLocalizedStringTransform
     /**
      * Transforms a normalized format into a localized money string.
      *
-     * @param number $value Normalized number
+     * @param int|float $value Normalized number
      *
      * @return string Localized money string.
      *
@@ -70,7 +70,7 @@ class MoneyToLocalizedStringTransformer extends NumberToLocalizedStringTransform
      *
      * @param string $value Localized money string
      *
-     * @return number Normalized number
+     * @return int|float Normalized number
      *
      * @throws TransformationFailedException If the given value is not a string
      *                                       or if the value can not be transformed.