+++ src/Symfony/Component/Form/FormTypeInterface.php
@@ -43,9 +43,9 @@ interface FormTypeInterface
      *
      * @see FormTypeExtensionInterface::buildView()
      *
-     * @param FormView $view    The view
-     * @param FormInterface     $form    The form
-     * @param array             $options The options
+     * @param FormView      $view    The view
+     * @param FormInterface $form    The form
+     * @param array         $options The options
      */
     public function buildView(FormView $view, FormInterface $form, array $options);
 
@@ -62,9 +62,9 @@ interface FormTypeInterface
      *
      * @see FormTypeExtensionInterface::finishView()
      *
-     * @param FormView $view    The view
-     * @param FormInterface     $form    The form
-     * @param array             $options The options
+     * @param FormView      $view    The view
+     * @param FormInterface $form    The form
+     * @param array         $options The options
      */
     public function finishView(FormView $view, FormInterface $form, array $options);