+++ src/Symfony/Component/Translation/Dumper/XliffFileDumper.php
@@ -21,6 +21,25 @@ use Symfony\Component\Translation\MessageCatalogue;
 class XliffFileDumper extends FileDumper
 {
     /**
+     * @var string
+     */
+    private $defaultLocale;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function dump(MessageCatalogue $messages, $options = array())
+    {
+        if (array_key_exists('default_locale', $options)) {
+            $this->defaultLocale = $options['default_locale'];
+        } else {
+            $this->defaultLocale = \Locale::getDefault();
+        }
+
+        parent::dump($messages, $options);
+    }
+
+    /**
      * {@inheritdoc}
      */
     protected function format(MessageCatalogue $messages, $domain)
@@ -33,7 +52,8 @@ class XliffFileDumper extends FileDumper
         $xliff->setAttribute('xmlns', 'urn:oasis:names:tc:xliff:document:1.2');
 
         $xliffFile = $xliff->appendChild($dom->createElement('file'));
-        $xliffFile->setAttribute('source-language', $messages->getLocale());
+        $xliffFile->setAttribute('source-language', str_replace('_', '-', $this->defaultLocale));
+        $xliffFile->setAttribute('target-language', str_replace('_', '-', $messages->getLocale()));
         $xliffFile->setAttribute('datatype', 'plaintext');
         $xliffFile->setAttribute('original', 'file.ext');